requirements.txt
bcrypt==4.0.1
Django==4.2.2
pandas==2.0.3
Pillow==10.0.0
pymongo==4.4.0